+++ piecrust/tests/crossover.rs
@@ -4,21 +4,21 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-use piecrust::{module_bytecode, Error, ModuleData, SessionData, VM};
-use piecrust_uplink::ModuleId;
+use piecrust::{contract_bytecode, ContractData, Error, SessionData, VM};
+use piecrust_uplink::ContractId;
 
 const OWNER: [u8; 32] = [0u8; 32];
 
-const CROSSOVER_ONE: ModuleId = {
+const CROSSOVER_ONE: ContractId = {
     let mut bytes = [0; 32];
     bytes[0] = 0x01;
-    ModuleId::from_bytes(bytes)
+    ContractId::from_bytes(bytes)
 };
 
-const CROSSOVER_TWO: ModuleId = {
+const CROSSOVER_TWO: ContractId = {
     let mut bytes = [0; 32];
     bytes[0] = 0x02;
-    ModuleId::from_bytes(bytes)
+    ContractId::from_bytes(bytes)
 };
 
 #[test]
@@ -29,12 +29,12 @@ fn crossover() -> Result<(), Error> {
     session.set_point_limit(u64::MAX / 100);
 
     session.deploy(
-        module_bytecode!("crossover"),
-        ModuleData::builder(OWNER).module_id(CROSSOVER_ONE),
+        contract_bytecode!("crossover"),
+        ContractData::builder(OWNER).contract_id(CROSSOVER_ONE),
     )?;
     session.deploy(
-        module_bytecode!("crossover"),
-        ModuleData::builder(OWNER).module_id(CROSSOVER_TWO),
+        contract_bytecode!("crossover"),
+        ContractData::builder(OWNER).contract_id(CROSSOVER_TWO),
     )?;
 
     // These value should not be set to `INITIAL_VALUE` in the contract.