+++ piecrust/tests/persistence.rs
@@ -4,7 +4,7 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-use piecrust::{module_bytecode, Error, ModuleData, SessionData, VM};
+use piecrust::{contract_bytecode, ContractData, Error, SessionData, VM};
 
 const OWNER: [u8; 32] = [0u8; 32];
 
@@ -18,10 +18,12 @@ fn session_commits_persistence() -> Result<(), Error> {
     let commit_1;
     {
         let mut session = vm.session(SessionData::builder())?;
-        id_1 = session
-            .deploy(module_bytecode!("counter"), ModuleData::builder(OWNER))?;
+        id_1 = session.deploy(
+            contract_bytecode!("counter"),
+            ContractData::builder(OWNER),
+        )?;
         id_2 = session
-            .deploy(module_bytecode!("box"), ModuleData::builder(OWNER))?;
+            .deploy(contract_bytecode!("box"), ContractData::builder(OWNER))?;
 
         session.call::<(), ()>(id_1, "increment", &())?;
         session.call::<i16, ()>(id_2, "set", &0x11)?;
@@ -51,7 +53,7 @@ fn session_commits_persistence() -> Result<(), Error> {
         let vm2 = VM::new(vm.root_dir())?;
         let mut session = vm2.session(SessionData::builder().base(commit_1))?;
 
-        // check if both modules' state was restored
+        // check if both contracts' state was restored
         assert_eq!(session.call::<(), i64>(id_1, "read_value", &())?, 0xfd);
         assert_eq!(
             session.call::<_, Option<i16>>(id_2, "get", &())?,
@@ -63,7 +65,7 @@ fn session_commits_persistence() -> Result<(), Error> {
         let vm3 = VM::new(vm.root_dir())?;
         let mut session = vm3.session(SessionData::builder().base(commit_2))?;
 
-        // check if both modules' state was restored
+        // check if both contracts' state was restored
         assert_eq!(session.call::<(), i64>(id_1, "read_value", &())?, 0xfe);
         assert_eq!(
             session.call::<_, Option<i16>>(id_2, "get", &())?,
@@ -74,13 +76,13 @@ fn session_commits_persistence() -> Result<(), Error> {
 }
 
 #[test]
-fn modules_persistence() -> Result<(), Error> {
+fn contracts_persistence() -> Result<(), Error> {
     let vm = VM::ephemeral()?;
     let mut session = vm.session(SessionData::builder())?;
     let id_1 = session
-        .deploy(module_bytecode!("counter"), ModuleData::builder(OWNER))?;
-    let id_2 =
-        session.deploy(module_bytecode!("box"), ModuleData::builder(OWNER))?;
+        .deploy(contract_bytecode!("counter"), ContractData::builder(OWNER))?;
+    let id_2 = session
+        .deploy(contract_bytecode!("box"), ContractData::builder(OWNER))?;
 
     session.call::<(), ()>(id_1, "increment", &())?;
     session.call::<i16, ()>(id_2, "set", &0x11)?;
@@ -95,7 +97,7 @@ fn modules_persistence() -> Result<(), Error> {
     let vm2 = VM::new(vm.root_dir())?;
     let mut session2 = vm2.session(SessionData::builder().base(commit_1))?;
 
-    // check if both modules' state was restored
+    // check if both contracts' state was restored
     assert_eq!(session2.call::<(), i64>(id_1, "read_value", &())?, 0xfd);
     assert_eq!(
         session2.call::<_, Option<i16>>(id_2, "get", &())?,